[untested] lanza excepciones desde la base de datos

+++ src/basededatos.cpp
@@ -1,5 +1,7 @@
 #include "basededatos.h"
 
+using std::string;
+
 BaseDeDatos::BaseDeDatos(){
 	Logger logger;
 	// Para optimizar RocksDB.
@@ -21,9 +23,9 @@ BaseDeDatos::~BaseDeDatos(){
 	logger.loggear("Se elimino la Base de Datos.", INFO);
 }
 
-bool BaseDeDatos::put(std::string key, std::string value) {
+bool BaseDeDatos::put(string key, string value) {
 	Logger logger;
-	std::string valor;
+	string valor;
 	Status s = db->Get(ReadOptions(), key, &valor);
 	if(s.IsNotFound()){
 		s = db->Put(WriteOptions(), key, value);
@@ -31,50 +33,59 @@ bool BaseDeDatos::put(std::string key, std::string value) {
 			logger.loggear("Se inserto satisfactoriamente la key (" + key + ") con el dato (" + value + ") en la base de datos." , TRACE);
 			return true;
 		} else {
-			return false;
+			logger.loggear("No se pudo insertar " + key + " en la base de datos", ERROR);
+			throw runtime_error("No se pudo insertar " + key + " en la base de datos");
 		}
-	} else {
-		logger.loggear("Se quiso insertar la key (" + key + ") con el dato (" + value + ") en la base de datos, pero ya se encontraba una key igual." , TRACE);
-//		std::cout << "La clave " << key << " ya existe, por favor use el metodo modify" << std::endl;
 	}
+	logger.loggear("Se quiso insertar la key (" + key + ") con el dato (" + value + ") en la base de datos, pero ya se encontraba una key igual." , TRACE);
 	return false;
-
 }
 
-std::string BaseDeDatos::get(std::string key) {
+string BaseDeDatos::get(string key) {
 	Logger logger;
-	std::string value;
+	string value;
 	Status s = db->Get(ReadOptions(), key, &value);
-//	if (!s.ok()) std::cout << "Get: " << s.ToString() << std::endl;
 	if (s.IsNotFound()) {
 		logger.loggear("No se encontro la clave (" + key + ") en la base de datos." , TRACE);
-//		std::cout << "No se encontro la clave " << key << " en la base de datos." << std::endl;
-		return "";
-	} else {
-		logger.loggear("Se obtuvo el valor: (" + value + ") de la base de datos, correspondiente a la clave (" + key + ")." , TRACE);
-		return value;
+		throw KeyNotFound("get: " + key);
+	} else if (not s.ok()) {
+		logger.loggear("No se pudo obtener la clave " + key + " de la base de datos", ERROR);
+		throw runtime_error("No se pudo obtener la clave " + key + " de la base de datos");
 	}
+	logger.loggear("Se obtuvo el valor: (" + value + ") de la base de datos, correspondiente a la clave (" + key + ")." , TRACE);
+	return value;
 }
 
-void BaseDeDatos::erase(std::string key) {
+void BaseDeDatos::erase(string key) {
 	Logger logger;
-	Status s = db->Delete(WriteOptions(), key);
+	string value;
+	Status s = db->Get(ReadOptions(), key, &value);
+	if (s.IsNotFound()) {
+		logger.loggear("Se quiso borrar la clave " + key + " inexistente en la base de datos", WARN);
+		throw KeyNotFound("erase: " + key);
+	}
+	s = db->Delete(WriteOptions(), key);
+	if (!s.ok()) {
+		std::cout << "Erase: " << s.ToString() << std::endl;
+		logger.loggear("No se pudo borrar la clave " + key + " de la base de datos", ERROR);
+		throw runtime_error("erase " + key);
+	}
 	logger.loggear("Se borro la clave (" + key + ") de la base de datos." , TRACE);
-//	if (!s.ok()) std::cout << "Erase: " << s.ToString() << std::endl;
 }
 
-void BaseDeDatos::modify(std::string key, std::string value) {
+void BaseDeDatos::modify(string key, string value) {
 	Logger logger;
-	std::string valor;
+	string valor;
 	Status s = db->Get(ReadOptions(), key, &valor);
-//	if(!s.ok()) std::cout << s.ToString() << std::endl;
 	if(s.IsNotFound()) {
-		logger.loggear("No se encontra la clave (" + key + ") en la base de dattos. Se intentaba modificar el valor (" + valor + ")." , TRACE);
-//		std::cout << "No se encuentra la key " << key << ", por favor use el metodo put." << std::endl;
-	} else {
-		s = db->Put(WriteOptions(), key, value);
-		logger.loggear("Se modifico la clave (" + key + "). Se cambio el valor (" + valor + ") por el valor: value " + value + "." , TRACE);
+		logger.loggear("No se encontra la clave (" + key + ") en la base de dattos. Se intentaba modificar el valor (" + valor + ")." , WARN);
+		throw KeyNotFound("modifiy: " + key);
+	} else if (not s.ok()){
+		logger.loggear("No se pudo modificar el valor en la clave " + key + " de la base de datos", ERROR);
+		throw new runtime_error("modify: " + key);
 	}
+	s = db->Put(WriteOptions(), key, value);
+	logger.loggear("Se modifico la clave (" + key + "). Se cambio el valor (" + valor + ") por el valor: value " + value + "." , TRACE);
 }
 
 void BaseDeDatos::deleteBD() {
